backend/requirements.txt
fastapi
uvicorn
websockets
anthropic
python-dotenv